backend/requirements.txt
asgiref==3.11.0
coverage==7.13.2
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
PyJWT==2.11.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytz==2025.2
sqlparse==0.5.5
